+++ src/typings/pouch.d.ts
@@ -1,5 +1,7 @@
 /**
  * this file contains typings that are pouchdb-specific
+ * most of it is copied from @types/pouchdb
+ * because it is outdated and strange
  */
 
 export interface PouchReplicationOptions {
@@ -34,5 +36,46 @@ export interface PouchSettings {
 
 export declare class PouchDB {
     constructor(name: string, options: { adapter: string });
-    info(): any;
+    info(): Promise<any>;
+
+    allDocs(options?: any): Promise<any>;
+    bulkDocs(
+        docs: Array<any>,
+        options?: any
+    ): Promise<any>;
+    compact(options?: any): Promise<any>;
+    destroy(options?: any): Promise<void>;
+    get(
+        docId: string,
+        options?: any
+    ): Promise<any>;
+    put(
+        doc: any,
+        options?: any,
+    ): Promise<any>;
+    remove(
+        doc: any | string,
+        options?: any,
+    ): Promise<any>;
+    changes(options?: PouchReplicationOptions): any;
+    close(): Promise<void>;
+    putAttachment(
+        docId: string,
+        attachmentId: string,
+        rev: string,
+        attachment: any,
+        type: string
+    ): Promise<any>;
+    getAttachment(
+        docId: string,
+        attachmentId: string,
+        options: { rev?: string },
+    ): Promise<any>;
+    removeAttachment(
+        docId: string,
+        attachmentId: string,
+        rev: string
+    ): Promise<void>;
+    bulkGet(options?: any): Promise<any>;
+    revsDiff(diff: any): Promise<any>;
 }